changed wifi upload and download speeds to float, and created seeds for each cafe

+++ db/migrate/20190528045312_create_wifi_speeds.rb
@@ -1,8 +1,8 @@
 class CreateWifiSpeeds < ActiveRecord::Migration[5.2]
   def change
     create_table :wifi_speeds do |t|
-      t.integer :upload_speed
-      t.integer :download_speed
+      t.float :upload_speed
+      t.float :download_speed
       t.integer :ping
 
       t.timestamps
